# pulse-query-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hospital Analytics Platform",
    "description": "Backend API for the Hospital Analytics Platform.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/dashboards/": {
      "get": {
        "tags": ["dashboards"],
        "summary": "List Dashboards",
        "description": "List all dashboards owned by the current user.",
        "operationId": "list_dashboards_api_v1_dashboards__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DashboardResponse"
                  },
                  "type": "array",
                  "title": "Response List Dashboards Api V1 Dashboards  Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": ["dashboards"],
        "summary": "Create Dashboard",
        "description": "Create a new empty dashboard.",
        "operationId": "create_dashboard_api_v1_dashboards__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/dashboards/{dashboard_id}/clone": {
      "post": {
        "tags": ["dashboards"],
        "summary": "Clone Dashboard",
        "description": "Creates a deep copy of an existing dashboard and all its widgets.\nThe new dashboard will have the name \"Copy of {original_name}\".\n\nArgs:\n    dashboard_id (UUID): The ID of the dashboard to clone.\n    current_user (User): Authenticated user (must own the dashboard).\n    db (AsyncSession): Database session.\n\nReturns:\n    DashboardResponse: The newly created dashboard populated with cloned widgets.",
        "operationId": "clone_dashboard_api_v1_dashboards__dashboard_id__clone_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/restore-defaults": {
      "post": {
        "tags": ["dashboards"],
        "summary": "Restore Default Dashboard",
        "description": "Re-creates the standard 'Hospital Command Center' dashboard.\nIf the dashboard already exists, creates a copy with a suffix to prevent data loss.\nPopulates the dashboard with all currently active templates in the registry.",
        "operationId": "restore_default_dashboard_api_v1_dashboards_restore_defaults_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/dashboards/{dashboard_id}": {
      "get": {
        "tags": ["dashboards"],
        "summary": "Get Dashboard",
        "description": "Get a specific dashboard details.",
        "operationId": "get_dashboard_api_v1_dashboards__dashboard_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["dashboards"],
        "summary": "Update Dashboard",
        "description": "Rename a dashboard.",
        "operationId": "update_dashboard_api_v1_dashboards__dashboard_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["dashboards"],
        "summary": "Delete Dashboard",
        "description": "Delete dashboard (and cascades to widgets).",
        "operationId": "delete_dashboard_api_v1_dashboards__dashboard_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/{dashboard_id}/widgets": {
      "post": {
        "tags": ["dashboards"],
        "summary": "Create Widget",
        "description": "Add a widget to the dashboard with Dry-Run Validation.",
        "operationId": "create_widget_api_v1_dashboards__dashboard_id__widgets_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/WidgetCreateSql"
                  },
                  {
                    "$ref": "#/components/schemas/WidgetCreateHttp"
                  },
                  {
                    "$ref": "#/components/schemas/WidgetCreateText"
                  }
                ],
                "discriminator": {
                  "propertyName": "type",
                  "mapping": {
                    "SQL": "#/components/schemas/WidgetCreateSql",
                    "HTTP": "#/components/schemas/WidgetCreateHttp",
                    "TEXT": "#/components/schemas/WidgetCreateText"
                  }
                },
                "title": "Widget In"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/widgets/{widget_id}": {
      "put": {
        "tags": ["dashboards"],
        "summary": "Update Widget",
        "description": "Update widget configuration (e.g., resize, change query) with Dry-Run.",
        "operationId": "update_widget_api_v1_dashboards_widgets__widget_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "widget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Widget Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WidgetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WidgetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["dashboards"],
        "summary": "Delete Widget",
        "description": "Delete a widget owned by the current user.",
        "operationId": "delete_widget_api_v1_dashboards_widgets__widget_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "widget_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Widget Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/{dashboard_id}/reorder": {
      "post": {
        "tags": ["dashboards"],
        "summary": "Reorder Widgets",
        "description": "Bulk update widget positions and groups.\nUsed for Drag-and-Drop persistence.",
        "operationId": "reorder_widgets_api_v1_dashboards__dashboard_id__reorder_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "dashboard_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Dashboard Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WidgetReorderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reorder Widgets Api V1 Dashboards  Dashboard Id  Reorder Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "DashboardCreate",
        "description": "Payload for creating a Dashboard."
      },
      "DashboardResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "widgets": {
            "items": {
              "$ref": "#/components/schemas/WidgetResponse"
            },
            "type": "array",
            "title": "Widgets",
            "default": []
          }
        },
        "type": "object",
        "required": ["name", "id", "owner_id"],
        "title": "DashboardResponse",
        "description": "API Response model for a Dashboard.\nIncludes the nested list of Widgets."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TextConfig": {
        "properties": {
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X",
            "description": "Grid X position",
            "default": 0
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y",
            "description": "Grid Y position",
            "default": 0
          },
          "w": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "W",
            "description": "Grid Width (cols)",
            "default": 6
          },
          "h": {
            "type": "integer",
            "minimum": 1.0,
            "title": "H",
            "description": "Grid Height (rows)",
            "default": 4
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Markdown text content."
          }
        },
        "type": "object",
        "required": ["content"],
        "title": "TextConfig",
        "description": "Configuration schema for Static Text/Markdown widgets.\n\nAttributes:\n    content (str): The static text content (Markdown supported)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
      "WidgetCreateHttp": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "type": {
            "type": "string",
            "const": "HTTP",
            "title": "Type"
          },
          "visualization": {
            "type": "string",
            "title": "Visualization",
            "description": "Visualization ID (e.g. 'table', 'metric')"
          },
          "config": {
            "$ref": "#/components/schemas/HttpConfig",
            "description": "HTTP-specific configuration"
          }
        },
        "type": "object",
        "required": ["title", "type", "visualization", "config"],
        "title": "WidgetCreateHttp",
        "description": "Payload for creating an HTTP Widget.\nStrictly enforces `type=\"HTTP\"` and `config` matching HttpConfig."
      },
      "WidgetCreateSql": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "type": {
            "type": "string",
            "const": "SQL",
            "title": "Type"
          },
          "visualization": {
            "type": "string",
            "title": "Visualization",
            "description": "Visualization ID (e.g. 'table', 'bar_chart')"
          },
          "config": {
            "$ref": "#/components/schemas/SqlConfig",
            "description": "SQL-specific configuration"
          }
        },
        "type": "object",
        "required": ["title", "type", "visualization", "config"],
        "title": "WidgetCreateSql",
        "description": "Payload for creating a SQL Widget.\nStrictly enforces `type=\"SQL\"` and `config` matching SqlConfig."
      },
      "WidgetCreateText": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title"
          },
          "type": {
            "type": "string",
            "const": "TEXT",
            "title": "Type"
          },
          "visualization": {
            "type": "string",
            "const": "markdown",
            "title": "Visualization",
            "description": "Fixed visualization ID for text.",
            "default": "markdown"
          },
          "config": {
            "$ref": "#/components/schemas/TextConfig",
            "description": "Text-specific configuration"
          }
        },
        "type": "object",
        "required": ["title", "type", "config"],
        "title": "WidgetCreateText",
        "description": "Payload for creating a Text Widget.\nStrictly enforces `type=\"TEXT\"` and `config` matching TextConfig."
      },
...
      "WidgetReorderRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WidgetReorderItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": ["items"],
        "title": "WidgetReorderRequest",
        "description": "Bulk update payload for persisting drag-and-drop results."
      },
      "WidgetResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "visualization": {
            "type": "string",
            "title": "Visualization"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "dashboard_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dashboard Id"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "visualization",
          "config",
          "id",
          "dashboard_id"
        ],
        "title": "WidgetResponse",
        "description": "API Response model for a persisted Widget."
      },
      "WidgetUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "visualization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visualization"
          },
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          }
        },
        "type": "object",
        "title": "WidgetUpdate",
        "description": "Payload for updating a widget.\nNote: Polymorphism is relaxed on Updates because 'type' is often immutable or missing\nin partial PATCH requests. We allow a loose dict for config, but individual fields\nare validated if matched.\n\nAttributes:\n    title (Optional[str]): New title.\n    visualization (Optional[str]): New visualization type.\n    config (Optional[Dict]): Partial configuration updates."
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}